+++ starsky/starsky/clientapp/src/components/archive-sidebar-label-edit-add-overwrite.tsx
@@ -86,6 +86,11 @@ const ArchiveSidebarLabelEditAddOverwrite: React.FunctionComponent = () => {
       setIsLoading(false);
       setInputEnabled(true);
 
+      // clear search cache
+      var searchTag = new URLPath().StringToIUrl(history.location.search).t;
+      if (!searchTag) return;
+      FetchPost(new UrlQuery().UrlSearchRemoveCacheApi(), `t=${searchTag}`);
+
     }).catch(() => {
       // loading + update button
       setIsLoading(false);