+++ starsky/starsky/clientapp/src/components/archive-sidebar-label-edit-search-replace.tsx
@@ -100,6 +100,11 @@ const ArchiveSidebarLabelEditSearchReplace: React.FunctionComponent = () => {
           setIsLoading(false);
           setInputEnabled(true);
 
+          // clear search cache * when you refresh the search page this is needed to display the correct labels
+          var searchTag = new URLPath().StringToIUrl(history.location.search).t;
+          if (!searchTag) return;
+          FetchPost(new UrlQuery().UrlSearchRemoveCacheApi(), `t=${searchTag}`);
+
         }).catch(() => {
           // loading + update button
           setIsLoading(false);