+++ starsky/starsky/clientapp/src/containers/detailview.tsx
@@ -118,8 +118,8 @@ const DetailView: React.FC<IDetailView> = () => {
     if (!history.location) return;
     if (new Keyboard().isInForm(event)) return;
 
-    var url = isSearchQuery ? new URLPath().Search(history.location.search) :
-      new URLPath().updateFilePath(history.location.search, state.fileIndexItem.parentDirectory);
+    var url = isSearchQuery ? new UrlQuery().HashSearchPage(history.location.search) :
+      new UrlQuery().updateFilePathHash(history.location.search, state.fileIndexItem.parentDirectory);
 
     history.navigate(url, {
       state: {