+++ starsky/starsky/clientapp/src/containers/login.tsx
@@ -3,10 +3,10 @@ import Button from '../components/Button';
 import useGlobalSettings from '../hooks/use-global-settings';
 import useLocation from '../hooks/use-location';
 import BrowserDetect from '../shared/browser-detect';
+import { DocumentTitle } from '../shared/document-title';
 import FetchGet from '../shared/fetch-get';
 import FetchPost from '../shared/fetch-post';
 import { Language } from '../shared/language';
-import { URLPath } from '../shared/url-path';
 import { UrlQuery } from '../shared/url-query';
 import { validateLoginForm } from '../shared/validate-login-form';
 
@@ -48,6 +48,7 @@ const Login: React.FC<ILoginProps> = () => {
   useEffect(() => {
     FetchGet(new UrlQuery().UrlAccountStatus()).then((status) => {
       setLogin(status.statusCode === 401);
+      new DocumentTitle().SetDocumentTitlePrefix(status.statusCode === 401 ? MessageLogin : MessageLogout);
       // to help new users find the register screen
       if (status.statusCode === 406 && history.location.search.indexOf(new UrlQuery().UrlAccountRegister()) === -1) {
         history.navigate(new UrlQuery().UrlAccountRegister(), { replace: true });
@@ -68,7 +69,7 @@ const Login: React.FC<ILoginProps> = () => {
       }
       else {
         // redirect
-        var returnUrl = new URLPath().GetReturnUrl(history.location.search);
+        var returnUrl = new UrlQuery().GetReturnUrl(history.location.search);
         history.navigate(returnUrl, { replace: true });
       }
     } catch (err) {
@@ -164,7 +165,7 @@ const Login: React.FC<ILoginProps> = () => {
             <form className="content--login-form">
               <div className="content--error-true">{LogoutWarning}</div>
               <a className="btn btn--default" href={new UrlQuery().UrlLogoutPage()}>{MessageLogout}</a>
-              <a className="btn btn--info" href={new UrlQuery().UrlHomePage()}>{MessageStayLoggedIn}</a>
+              <a className="btn btn--info" href={new UrlQuery().GetReturnUrl(history.location.search)}>{MessageStayLoggedIn}</a>
             </form>
           </div>
         </>