+++ starsky/starsky/clientapp/src/shared/url-query.ts
@@ -9,8 +9,27 @@ export class UrlQuery {
     return document.location.pathname.indexOf(this.prefix) === -1 ? "/" : `${this.prefix}/`;
   }
 
+  public GetReturnUrl(locationHash: string): string {
+    // ?ReturnUrl=%2F
+    let hash = new URLPath().RemovePrefixUrl(locationHash);
+    let search = new URLSearchParams(hash);
+    let getReturnUrl = search.get("ReturnUrl");
+    var starskyPathIndex = document.location.pathname.indexOf(this.prefix);
+    if (!getReturnUrl) return starskyPathIndex === -1
+      ? `/${new URLPath().AddPrefixUrl("f=/")}` : `${this.prefix}/${new URLPath().AddPrefixUrl("f=/")}`;
+    return starskyPathIndex === -1 ? getReturnUrl : `${this.prefix}/${getReturnUrl}`;
+  }
+
   public UrlSearchPage(t: string): string {
     return document.location.pathname.indexOf(this.prefix) === -1 ? `/search?t=${t}` : `${this.prefix}/search?t=${t}`;
+  }
+
+  /**
+   * Search path based on Location Hash
+    */
+  public HashSearchPage(historyLocationHash: string): string {
+    var url = new URLPath().StringToIUrl(historyLocationHash);
+    return document.location.pathname.indexOf(this.prefix) === -1 ? `/search${new URLPath().IUrlToString(url)}` : `${this.prefix}/search${new URLPath().IUrlToString(url)}`;
   }
 
   public UrlTrashPage(): string {
@@ -28,6 +47,7 @@ export class UrlQuery {
   public UrlLogoutPage(): string {
     return document.location.pathname.indexOf(this.prefix) === -1 ? `/account/logout` : `${this.prefix}/account/logout`;
   }
+
 
   private urlReplacePath(input: string): string {
     let output = input.replace("#", "");
@@ -69,6 +89,20 @@ export class UrlQuery {
   }
 
   /**
+   * Keep colorClass in URL
+   */
+  public updateFilePathHash(historyLocationHash: string, toUpdateFilePath: string, clearTSearchQuery?: boolean): string {
+    var url = new URLPath().StringToIUrl(historyLocationHash);
+    url.f = toUpdateFilePath;
+    // when browsing to a parent folder from a detailview item
+    if (clearTSearchQuery) {
+      delete url.t;
+      delete url.p;
+    }
+    return document.location.pathname.indexOf(this.prefix) === -1 ? `/${new URLPath().IUrlToString(url)}` : `${this.prefix}/${new URLPath().IUrlToString(url)}`;
+  }
+
+  /**
    * Used with localisation hash
    */
   public UrlQueryServerApi = (historyLocationHash: string): string => {
@@ -102,7 +136,7 @@ export class UrlQuery {
    * Get Direct api/index with IUrl
    */
   public UrlIndexServerApiPath = (path: string): string => {
-    return this.prefix + "/api/index" + path
+    return this.prefix + "/api/index?f=" + path
   }
 
   /**